apps/trade_hunter/pyproject.toml
[project]
name = "trade_hunter"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
  "typer>=0.12",
  "rich>=13",
  "python-dotenv>=1.0",
  "pandas>=2.2",
  "openpyxl>=3.1",
  "httpx>=0.27",
]

[project.scripts]
trade_hunter = "trade_hunter.cli:app"

[dependency-groups]
dev = ["pytest>=8"]


[build-system]
requires = ["hatchling>=1.24"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/trade_hunter"]
